+++ ChangeLog.md
@@ -1,5 +1,16 @@
 # Wasat ChangeLog
 
+## Unreleased
+
+**Released: WiP**
+
+- Added `--verify-mode` to the CLI of the library so the certificate
+  verification mode can be tested.
+  ([#39](https://github.com/davep/wasat/pull/39))
+- Added `hybrid` certificate verification mode to `Client` to combine system
+  CA validation with TOFU fallback.
+  ([#39](https://github.com/davep/wasat/pull/39))
+
 ## v1.0.1
 
 **Released: 2026-07-21**